# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Context-Aware PQC Risk Assessment API
  version: 0.1.0
  license:
    name: Internal Capstone Project
    url: https://github.com/ehdgks0627/crypto-scanner
  description: |
    API contract for the PQC Risk Assessment System.

    `docs/api/openapi.yaml` is the design source of truth. The Django Ninja
    implementation must expose an equivalent `/api/openapi.json`.
servers:
  - url: /api
security:
  - ApiToken: []
...
paths:
...
  /snapshots:
    get:
      tags: [Snapshots]
      summary: List CBOM snapshots
      operationId: listSnapshots
      parameters:
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Snapshot list
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CbomSnapshotPage'
              examples:
                default:
                  externalValue: ./examples/snapshot-list.json
        '400':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Snapshots]
      summary: Get snapshot metadata
      operationId: getSnapshot
      responses:
        '200':
          description: Snapshot metadata
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CbomSnapshot'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{id}/export:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Snapshots]
      summary: Export snapshot as CycloneDX CBOM JSON
      operationId: exportSnapshot
      parameters:
        - name: pretty
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: CycloneDX CBOM JSON
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Content-Disposition:
              $ref: '#/components/headers/ContentDisposition'
          content:
            application/vnd.cyclonedx+json:
              schema:
                type: object
                additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{id}/diff:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Snapshots]
      summary: Diff two snapshots
      operationId: diffSnapshots
      parameters:
        - name: other
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Snapshot diff
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CbomDiff'
              examples:
                default:
                  externalValue: ./examples/snapshot-diff.json
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{sid}/assets:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    get:
      tags: [Assets]
      summary: List assets in a snapshot
      operationId: listAssets
      parameters:
        - name: asset_class
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetClass'
        - name: asset_type
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetType'
        - name: target_id
          in: query
          schema:
            type: integer
        - name: min_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: max_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: tier
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RiskTier'
        - name: quantum_vulnerable
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Asset inventory
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListPage'
              examples:
                default:
                  externalValue: ./examples/asset-list.json
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /snapshots/{sid}/risks:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    get:
      tags: [Risk]
      summary: List risk scores for a snapshot
      operationId: listRisks
      parameters:
        - name: tier
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RiskTier'
        - name: min_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: max_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Risk score list
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScorePage'
              examples:
                default:
                  externalValue: ./examples/risk-list.json
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{sid}/risks/top:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    get:
      tags: [Risk]
      summary: List top-N risk scores
      operationId: listTopRisks
      parameters:
        - name: n
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Top risks
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskScorePage'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{sid}/recompute:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    post:
      tags: [Risk]
      summary: Recompute risk scores for a snapshot
      operationId: recomputeRisks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskRecomputeRequest'
      responses:
        '202':
          description: Recompute job accepted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
              examples:
                accepted:
                  externalValue: ./examples/recompute-job-accepted.json
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        default:
          $ref: '#/components/responses/Error'
...
  /snapshots/{sid}/migration-plan:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    get:
      tags: [Migration]
      summary: Get migration plan items for a snapshot
      operationId: getMigrationPlan
      parameters:
        - name: min_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: tier
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/RiskTier'
        - name: asset_type
          in: query
          style: form
          explode: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetType'
        - name: target_id
          in: query
          schema:
            type: integer
        - name: asset_ids
          in: query
          description: Comma-separated asset ids to include.
          style: form
          explode: false
          schema:
            type: array
            items:
              type: integer
            minItems: 1
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Migration plan
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlanPage'
              examples:
                default:
                  externalValue: ./examples/migration-plan.json
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        default:
          $ref: '#/components/responses/Error'
  /snapshots/{sid}/migration-plan/impact:
    parameters:
      - $ref: '#/components/parameters/SnapshotId'
    get:
      tags: [Migration]
      summary: Get migration impact analysis for selected assets
      operationId: getMigrationImpact
      parameters:
        - name: asset_ids
          in: query
          required: true
          description: Comma-separated asset ids.
          style: form
          explode: false
          schema:
            type: array
            items:
              type: integer
            minItems: 1
      responses:
        '200':
          description: Migration impact analysis
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationImpact'
              examples:
                default:
                  externalValue: ./examples/migration-impact.json
        '400':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-API-Token
    BootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
    AgentBearer:
      type: http
      scheme: bearer
  headers:
    XRequestId:
      description: Request trace id. If the client sends `X-Request-Id`, the server echoes it.
      schema:
        type: string
        format: uuid
...
    ContentDisposition:
      description: Download filename for exported files.
      schema:
        type: string
      example: attachment; filename="cbom-56.json"
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    SnapshotId:
      name: sid
      in: path
      required: true
      schema:
        type: integer
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: Field name, prefixed with `-` for descending order.
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              externalValue: ./examples/error-validation.json
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unprocessable:
      description: Request is semantically invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Required backend dependency is unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    JobKind:
      type: string
      enum: [scan_job, discovery, recompute]
...
    RiskTier:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW]
    AssetClass:
      type: string
      enum: [crypto, host, service, data]
    AssetType:
      type: string
      enum: [algorithm, certificate, key, protocol, keystore, device, service, data]
...
    ErrorResponse:
      type: object
      required: [error, message, details]
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
        details:
          type: object
          description: Machine-readable error details. Empty object when there are no details.
          additionalProperties: true
...
    PageBase:
      type: object
      required: [total, offset, limit]
      properties:
        total:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
...
    JobEnvelope:
      type: object
      required: [id, kind, resource, status, progress, started_at, cancel_requested_at, finished_at, result, error]
      properties:
        id:
          type: integer
          description: API-visible asynchronous job id, unique across all job kinds.
        kind:
          $ref: '#/components/schemas/JobKind'
        resource:
          $ref: '#/components/schemas/JobResource'
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobProgress'
          nullable: true
          description: Always present. Null while PENDING or when no progress can be reported.
        started_at:
          type: string
          format: date-time
          nullable: true
        cancel_requested_at:
          type: string
          format: date-time
          nullable: true
          description: Set when cancellation has been requested but a running worker has not finished cleanup yet.
        finished_at:
          type: string
          format: date-time
          nullable: true
        result:
          type: object
          allOf:
            - $ref: '#/components/schemas/JobResult'
          nullable: true
          example:
            snapshot_id: 56
        error:
          type: string
          nullable: true
...
    CbomSnapshot:
      type: object
      required:
        - id
        - scan_job_id
        - serial_number
        - asset_count
        - created_at
        - summary
        - validation_errors
      properties:
        id:
          type: integer
        scan_job_id:
          type: integer
          nullable: true
        serial_number:
          type: string
          example: urn:uuid:9f8e7d6c-0000-4000-8000-000000000056
        asset_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        summary:
          $ref: '#/components/schemas/SnapshotSummary'
        validation_errors:
          type: array
          items:
            type: object
            additionalProperties: true
          default: []
    SnapshotSummary:
      type: object
      properties:
        by_tier:
          type: object
          additionalProperties:
            type: integer
        by_type:
          type: object
          additionalProperties:
            type: integer
    CbomSnapshotPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/CbomSnapshot'
    CbomDiff:
      type: object
      required: [snapshot_a, snapshot_b, added, removed, modified, unchanged_count]
      properties:
        snapshot_a:
          type: integer
        snapshot_b:
          type: integer
        added:
          type: array
          items:
            $ref: '#/components/schemas/CbomDiffAsset'
        removed:
          type: array
          items:
            $ref: '#/components/schemas/CbomDiffAsset'
        modified:
          type: array
          items:
            $ref: '#/components/schemas/CbomDiffModifiedAsset'
        unchanged_count:
          type: integer
    CbomDiffAsset:
      type: object
      required: [bom_ref, type, name]
      properties:
        bom_ref:
          type: string
        type:
          type: string
        name:
          type: string
...
    AssetListPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AssetListItem'
...
    RiskScorePage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RiskScore'
...
    RiskRecomputeRequest:
      type: object
      required: [weights]
      properties:
        weights:
          $ref: '#/components/schemas/RiskWeightsInput'
        persist_weights_as_default:
          type: boolean
          description: Defaults to false when omitted.
...
    MigrationPlanPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/MigrationPlanItem'
    MigrationImpact:
      type: object
      required:
        - selected_count
        - hosts
        - services
        - cert_reissues
        - config_changes
        - key_regens
        - estimated_downtime_min
      properties:
        selected_count:
          type: integer
          minimum: 0
        hosts:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        cert_reissues:
          type: integer
          minimum: 0
        config_changes:
          type: integer
          minimum: 0
        key_regens:
          type: integer
          minimum: 0
        estimated_downtime_min:
          type: integer
          minimum: 0
...